Move to separate package security.jwt
