+++ solr/core/src/java/org/apache/solr/security/AuthenticationPlugin.java
@@ -95,11 +95,11 @@ public abstract class AuthenticationPlugin implements SolrInfoBean {
     }
   }
 
-  HttpServletRequest wrapWithPrincipal(HttpServletRequest request, Principal principal) {
+  protected HttpServletRequest wrapWithPrincipal(HttpServletRequest request, Principal principal) {
       return wrapWithPrincipal(request, principal, principal.getName());
   }
 
-  HttpServletRequest wrapWithPrincipal(HttpServletRequest request, Principal principal, String username) {
+  protected HttpServletRequest wrapWithPrincipal(HttpServletRequest request, Principal principal, String username) {
     return new HttpServletRequestWrapper(request) {
       @Override
       public Principal getUserPrincipal() {
